+++ blocklytest/src/androidTest/java/com/google/blockly/model/BlocklyEventTest.java
@@ -2,7 +2,6 @@ package com.google.blockly.model;
 
 import android.support.test.InstrumentationRegistry;
 
-import com.google.blockly.android.control.BlocklyEvent;
 import com.google.blockly.android.test.R;
 import com.google.blockly.utils.BlocklyXmlHelper;
